requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.4

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.9

# NLP
spacy==3.7.4

# File Processing
pypdf==4.0.0
python-docx==1.1.0

# AWS (optional for S3)
boto3==1.34.0

# Validation & Settings
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Utils
aiofiles==23.2.1
httpx==0.27.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Frontend
streamlit==1.31.0
requests==2.31.0
